added scenarios into the URL as search parameters
I've added the scenarios (and the showDifference boolean flag) into the URL as search parameters; this is rather than having them as state properties which are propagated through the app.

This is to enable the sharing (e.g. by email to clients, or on social media) of URL links which directly to not only a specific set of Charts, but the Charts for a specific Main Scenario (together with a specific Compare Scenario).

It does introduce one backwards incompatibility: scenario names MUST now be made URI-safe before going into config.js and scenarioTitles.js